Within Subjects Confidence Intervals.

within_ci returns lower and upper bounds of confidence intervals per factor for data with at least one within subject factor.

within_ci(.data, .subject, .value, .conf = 0.95, .ws_factors = c(),
  .bs_factors = c())

Arguments

.data:
Dataframe in long format

.subject:
Name of single column in .data containing participant identifiers

.value:
Name of single column in .data containing values of DV

.conf:
Confidence interval to be calculated

.ws_factors:
names of (multiple) columns in .data containing within-subject factors

.bs_factors:
names of (multiple) columns in .data containing between-subject factors

Value

Returns length of Confidence Intervals per factor

Examples

# One within factor, no between factors
WithinS_CI(.data = data_1w,
           .subject = subject,
           .value = DV,
           .ws_factors = within1)

#> Error in WithinS_CI(.data = data_1w, .subject = subject, .value = DV,     .ws_factors = within1): could not find function "WithinS_CI"

# Multiple within and between factors
WithinS_CI(.data = data_2w2b,
           .subject = subject,
           .value = DV,
           .ws_factors = c(within1, within2),
           .bs_factors = c(between1, between2))

#> Error in WithinS_CI(.data = data_2w2b, .subject = subject, .value = DV,     .ws_factors = c(within1, within2), .bs_factors = c(between1,         between2)): could not find function "WithinS_CI"

# Multiple within and between factors with graph
 data_2w2b %>%
   WithinS_CI(.data = .,
             .subject = subject,
             .value = DV,
             .ws_factors = c(within1, within2),
             .bs_factors = c(between1, between2)) %>%
  ggplot(aes(x = within1,
             y = sample_mean,
             group = within2)) +
  geom_bar(stat = "identity",
           position = position_dodge(width = 1),
           color = "black",
           aes(fill = within2)) +
  geom_errorbar(position = position_dodge(width = 1),
                width = .1,
                aes(ymin = sample_mean - CI,
                    ymax = sample_mean + CI)) +
  scale_fill_manual(values = c("white", "grey")) +
  labs(y = "DV",
       title = "Example bar plot - two within two between") +
  theme_classic() +
  facet_grid(between1 ~ between2)

#> Error in data_2w2b %>% WithinS_CI(.data = ., .subject = subject, .value = DV,     .ws_factors = c(within1, within2), .bs_factors = c(between1,         between2)) %>% ggplot(aes(x = within1, y = sample_mean,     group = within2)): could not find function "%>%"